Fix Errors#from_type use known_attributes
This commit replaces the use of #attributes.keys with #known_attributes
in Errors#from_array and Errors#from_hash. This is done so that when
ActiveResource grabs errors for attributes defined in the schema, but
don't yet exist as an attribute on the object, the error gets assigned
to the known attribute instead of the base object.