Example for collection parser is incorrect.
Either "self.collection_parser = PostParser" in the original example needs to be changed to refer to the subsequent PostCollection class (the path I chose) or the PostCollection class needs to be named PostParser. Bottom line is the class name of the class that inherits from ActiveResource::Collection needs to match the name specified as collection_parser.