Avoid using alias_method_chain, now deprecated in Active Support
The less than venerable `#alias_method_chain` has been deprecated in the
latest version of Active Support. Fortunately, we only use it in two places.

Replace with the longhand version using `alias_method` for forward
compatibility. Not using Module#prepend in order to maintain compatibility
with older versions of Ruby.