Upgrade ActiveResource for Rails 5
Rails 5 requires Ruby 2.2.2 or higher. Tests for 2.0 and below have been
failing for some time on this gem anyway.

Additionally this adds `activemodel-serializers-xml` since that was
extracted to a gem and removed in Rails 5.

I swapped the Rails in the gemfile to use the specified gems so we don't
have to require sprockets and sprockets-rails in it since this gem
doesn't even use those gems.