Exclude params from model's update.
If your API throws an exception when unexpected params are received, you
must have a way to stop sending them (even if you receive them in a get
request). This change introduces two new `ActiveResource::Base`'s class
methods: `update_except` and `update_except_primary_key`. This way when
updating a model, the params specified in `update_except` are not sent
to the API, and if the `update_except_primary_key` is set, the `id` is
not sent in the request's body.