The logger nows logs a synthetic '523' error. Fix 150
When connection can't even be established, the logger
gets a nil result; this used to raise an exception and
dump a lot of stack traces. This patch makes the logger
proceed without raising the exception; instead it logs
the following message:

  523 "ActiveResource connection error"

The 523 error code was chosen to match the closest
situation I can think of to describe the event (it's
a CloudFlare code), since really this error situation
occurs completely outside of the HTTP protocol anyway.