use UnknownResource when resource fails to create normally
When loading a resource, checking if the key responds to `to_sym` is
insufficient to ensure a valid resource is created. An integer like
1234 will not respond to `to_sym` but "1234" will. Attempting to
check const_defined? for :"1234" will result in a "NameError: wrong
constant name 1234"

Similarly, any string with special characters or spaces will error as
well.

Instead, this stops checking for symbols in Base#split_options and will
catch the NameError exception should it arise and create a resource
called UnkownResource instead which will hold the resource attributes.