Add `activerecord` integration test coverage
Follow-up to [#420][]

Background
---

The introduction of `ActiveResource::Base.coder` and the
`ActiveResource::Coder` class added support for encoding and decoding
instances of `Base`. The intent was to integrate with the Active Record
[.serialize][] method.

Problem
---

While the `Coder` class and `Serialization` module aim to comply with
Active Record, there's still potential for that compatibility to drift
of break without sufficient test coverage.

Proposal
---

This commit adds a test-only dependency on `activerecord` and `sqlite3`
so that the serialization test coverage can include database-level reads
and writes. The suite is configured to use the in-memory SQLite adapter
so that there are no need for migrations or external database
configuration.

[#420]: https://github.com/rails/activeresource/pull/420
[.serialize]: https://api.rubyonrails.org/classes/ActiveRecord/AttributeMethods/Serialization/ClassMethods.html#method-i-serialize