Remove Validations deprecation code
The `ActiveResource::Validations#from_json` method emits deprecation
warnings when handling error response payloads that resemble the
following:

```
{"errors":["First cannot be empty"]}

{"first":["cannot be empty"]}
```

The deprecation was originally introduced in [c643605][] (Feb 4, 2012).
This commit removes documentation mentions of backwards-compatibility
along with the conditional branches that emit the deprecation warnings.

The removal of the backwards support is a breaking change.

[c643605]: https://github.com/rails/activeresource/commit/c6436053af9b6e0f51e1f063a0898aba4738637d