Singletons: Fix bug that prevented callbacks
Prior to this commit, the `Singleton` module definitions for `create`,
`_update`, and `destroy` methods were re-written in a way that did not
integrate with callbacks defined by `Base` (through its inclusion of the
`Callbacks` concern).

This commit resolves that issue by passing a default positional `path`
argument to the `Base` implementations for `create`, `_update`, and
`destroy`. The `Singleton` module will redefine those methods with
singleton-specific default values when it's mix-in.

The `destroy` method is the only method that is part of the public API.
As to not introduce a breaking change to that interface, this commit
adds a `:call-seq:` comment to preserve its argument-less documentation.