Singleton: Inline `find_singleton` back into `find`
When the Singleton module was originally introduced in [#4][], there was
a suggestion about the implementation for `.find` that used the `:one`
modifier option with the `:from` keyword argument.

```ruby
class Account < ActiveResource::Base
  def self.find_singleton
    find :one, from: '/account.xml'
  end
end
```

This commit modifies the `Singleton.find` method by inlining the private
`find_singleton` method. By doing so, the implementation can invoke
`super`, and rely on the built-in `instantiate_record` call provided by
`Base`.

This reduces the amount of duplication involved in transforming a
response payload into a resource instance.

[#4]: https://github.com/rails/activeresource/pull/4
[#4]: https://github.com/rails/activeresource/pull/4#issuecomment-4697435