Prefix: restore variable for the sake of error logging
Resolve a 13 year old bug (according to a [comment on 7589c1a][]) by
re-declaring the `code` variable that was unintentionally removed in
[da74c4f][].

Prior to this commit, the following error is raised when setting
`.prefix = nil`:

```
NameError: undefined local variable or method `code' for class Person
```

However, the error should be a `NoMethodError` from calling
`String#gsub` on `nil`, rather than the `NameError` raised by the
missing variable. The `NoMethodError` can still be improved upon, but
this commit aims to restore the logging visibility first-and-foremost.

[comment on 7589c1a]: https://github.com/rails/activeresource/commit/7589c1a3c234ebb12880b81b1999b0cbe97b8402#r2381147
[da74c4f]: https://github.com/rails/activeresource/commit/da74c4ffcfcafdd36c5589948a5ea7eb55af60d0#diff-5d3e28bbfdc782e5e5501903dd1171cc83b1aef685cc5b0547f527ec2573f989R58