Deprecate writing to `Base#attributes`
The Base class defines an `attr_accessor :attributes` that is marked
with `# :nodoc:`. Technically, this means that any interaction with
`Base#attributes` or `Base#attributes=` is not part of the public
interface, and is free to be changed or removed without breaking the
public API.

However, given the project's age and the long-term period of time with
minimal changes to the public interface, this PR proposes that there be
a public deprecation of **writing** to the Hash instance returned by the
`Base#attributes` method.

The migration to `ActiveModel::Attributes` proposed in [#410][] will
involve changes to the `Base#attributes` method (due to
[ActiveModel::Attributes#attributes][]). Reading from the value returned
by `ActiveModel::Attributes#attributes` will remain the same, but
writing to that value will have no affect since the value is a Hash copy
created by [ActiveModel::AttributeSet#to_hash][], and not the Hash
instance used internally.

Similarly, `ActiveModel::Attributes` does not expose a corresponding
`#attributes=` method. By deprecating `#attributes=`, changes made that
incorporate `ActiveModel::Attributes` will not need to add an otherwise
unnecessary `#attributes=` implementation.

Once deprecated and part of a release cycle, the
`ActiveResource::AttributeSet` class can be removed, and the Active
Model migration's backwards compatibility burden can be reduced.

[#410]: https://github.com/rails/activeresource/pull/410
[ActiveModel::Attributes#attributes]: https://edgeapi.rubyonrails.org/classes/ActiveModel/Attributes.html#method-i-attributes
[ActiveModel::AttributeSet#to_hash]: https://github.com/rails/rails/blob/v8.1.1/activemodel/lib/active_model/attribute_set.rb#L36-L39