Assign `Net::HTTPRequest` to exceptions
With the changes made in [2ea5401][], the underlying
[Net::HTTPRequest][] instance is available to the connection throughout
the lifetime of the request. With that instance available, exceptions
raised during the request-response handshake can be constructed with the
`Net::HTTPRequest` instance.

Access to the `Net::HTTPRequest` from code that rescues from those
exceptions is  valuable:

* for debugging purposes
* for observability and logging
* for retrying requests, etc.

However, the exception classes are documented and therefore their
constructors are part of the public API. Inserting a `request`
positional argument is a breaking change. To account for that, this
commit includes deprecations for subclasses that call `super` from
within `#initialize` to instruct them on changing the signature to
account for the `Net::HTTPRequest` instance.

[2ea5401]: https://github.com/rails/activeresource/commit/2ea5401a56a8e0060de4d34ec2eb693ff60b84c6
[Net::HTTPRequest]: https://docs.ruby-lang.org/en/master/Net/HTTPRequest.html